refactor: move message structs and enums to msg.rs, temporarily disable chainID check
